+++ db/migrate/20190529210803_index_eve_alliance_corporations.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class IndexEveAllianceCorporations < ActiveRecord::Migration[6.0]
+  def change
+    add_index :eve_alliance_corporations, [:alliance_id, :corporation_id], unique: true, name: :index_eve_ac_on_alliance_id_and_corporation_id
+  end
+end