+++ db/migrate/20190529213042_int_to_bigint_loyalty_points.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class IntToBigintLoyaltyPoints < ActiveRecord::Migration[6.0]
+  def change
+    change_column :loyalty_points, :corporation_id, :bigint
+  end
+end