+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 * Show Alliance corporations
 * Migrate to Docker
 * Handle service unavailable (503) on sign in via EVE SSO
-* Rewrite working with etags
+* Rewrite working with etags and updating public info
+* Import eve types info
 
 ## v0.1.0